Move interface to base
