+++ packages/adapters/wallet-connect-v1-adapter/package.json
@@ -41,8 +41,7 @@
     "@walletconnect/client": "~1.7.7",
     "@walletconnect/types": "~1.7.7",
     "@web3auth/base": "^0.9.4",
-    "@web3auth/ethereum-provider": "^0.9.4",
-    "@web3auth/ui": "^0.9.4"
+    "@web3auth/ethereum-provider": "^0.9.4"
   },
   "lint-staged": {
     "!(*d).ts": [
