+++ packages/ui/src/networkSwitch.tsx
@@ -1,12 +1,11 @@
 /* eslint-disable class-methods-use-this */
 import "../css/network.css";
 
-import { CustomChainConfig } from "@web3auth/base";
+import { BaseNetworkSwitch, CustomChainConfig } from "@web3auth/base";
 import { render } from "react-dom";
 
 import AddNetwork from "./components/AddNetwork";
 import SwitchNetwork from "./components/SwitchNetwork";
-import { BaseNetworkSwitch } from "./interfaces";
 
 function createWrapper(): HTMLElement {
   const wrapper = document.createElement("section");
